third_party/WebKit/LayoutTests/svg/dom/set-title-html-namespace-on-SVG.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
  // svg != SVG
  var doc = document.implementation.createDocument("http://www.w3.org/2000/svg", "SVG", null);
  doc.documentElement.appendChild(doc.createElementNS("http://www.w3.org/1999/xhtml", "title"));
  doc.documentElement.lastChild.textContent = "bar";
  assert_equals(doc.title, "bar");

  doc.documentElement.appendChild(doc.createElementNS("http://www.w3.org/1999/xhtml", "title"));
  doc.documentElement.lastChild.textContent = "foo";
  assert_equals(doc.title, "bar");
}, "Test for setting title element in the html namespace on SVG element that is root.");
</script>
